WIP: expand trino columns
Add an extra optional query_as field to the column definition so that we
can override the way it's queried: otherwise sqlalchemy is going to
quote the whole thing as a single column name, which isn't correct, it
should actually be quoted per dotted segment, and we'll want to alias it
to the full dotted string.

Add some temporary logic into get_columns to inject a dummy column into
the output to assess how this will work; once happy we'll replace this
with dotted paths assembled using our type parser.